# backend/docs/swagger.yaml
basePath: /
definitions:
...
  kategoriModel.Kategori:
    properties:
      deskripsi:
        type: string
      id:
        type: string
      nama:
        type: string
      ref_kategori:
        $ref: "#/definitions/kategoriModel.RefKategori"
    type: object
...
  kategoriModel.KategoriReference:
    properties:
      deskripsi:
        type: string
      id:
        type: string
      nama:
        type: string
      ref_kategori_id:
        type: string
    type: object
...
  kategoriModel.RefKategori:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
...
  kategoriModel.Reference:
    properties:
      id:
        type: string
      nama:
        type: string
      values:
        items:
          $ref: "#/definitions/kategoriModel.KategoriReference"
        type: array
    type: object
...
  model.PaginationMeta:
    properties:
      next:
        type: string
      page:
        type: integer
      previous:
        type: string
      reference:
        $ref: "#/definitions/model.ReferenceResult"
      size:
        type: integer
      total_data:
        type: integer
    type: object
...
  model.PaginationResponseTest:
    properties:
      meta:
        $ref: "#/definitions/model.PaginationMeta"
      result:
        items:
          $ref: "#/definitions/produkModel.Produk"
        type: array
    type: object
  model.ReferenceResult:
    properties:
      data:
        items:
          $ref: "#/definitions/kategoriModel.Reference"
        type: array
      last:
        type: string
    type: object
  model.ResponseMessage:
    properties:
      message:
        type: string
      reference:
        $ref: "#/definitions/model.ReferenceResult"
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  produkModel.Produk:
    properties:
      detail:
        type: string
      foto:
        type: string
      harga:
        type: integer
      id:
        type: string
      kategori:
        $ref: "#/definitions/kategoriModel.Kategori"
      kategori_id:
        type: string
      nama:
        type: string
    type: object
...
  produkModel.ProdukUpdate:
    properties:
      detail:
        type: string
      foto:
        type: string
      harga:
        type: integer
      kategori_id:
        type: string
      nama:
        type: string
    type: object
...
  userModel.UserPatch:
    properties:
      email:
        type: string
      foto:
        type: string
      nama:
        type: string
      no_telp:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  userModel.UserShow:
    properties:
      email:
        type: string
      foto:
        type: string
      id:
        type: string
      nama:
        type: string
      no_telp:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: dokumentasi API Kavi Kasir
  title: KAVI Kasir API
  version: "1.0"
paths:
...
  /produk/{produk_id}:
    delete:
      consumes:
        - application/json
      description: Menghapus data produk
      parameters:
        - description: produk_id
          in: path
          name: produk_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Hapus produk
      tags:
        - produk
    get:
      consumes:
        - application/json
      description: ambil data produk
      parameters:
        - description: produk_id
          in: path
          name: produk_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/produkModel.Produk"
      summary: Ambil produk berdasarkan id
      tags:
        - produk
    put:
      consumes:
        - application/json
      description: Mengubah data produk
      parameters:
        - description: produk_id
          in: path
          name: produk_id
          required: true
          type: string
        - description: Field produk
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/produkModel.ProdukUpdate"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Ubah produk
      tags:
        - produk
  /produk/{user_id}:
    patch:
      consumes:
        - application/json
      description: Update data user
      parameters:
        - description: user_id
          in: path
          name: user_id
          required: true
          type: string
        - description: Field user
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/userModel.UserPatch"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/userModel.UserShow"
      summary: Update data user
      tags:
        - user
...
  /users:
    get:
      consumes:
        - application/json
      description: Ambil semua data user
      parameters:
        - description: page
          in: path
          name: page
          type: integer
        - description: size
          in: path
          name: size
          type: integer
        - description: keyword
          in: path
          name: keyword
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.PaginationResponseTest"
      summary: Ambil semua data user
      tags:
        - user
  /users/{user_id}:
    delete:
      consumes:
        - application/json
      description: Menghapus user
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Menghapus data user
      tags:
        - user
    get:
      consumes:
        - application/json
      description: Menampilkan detail user
      parameters:
        - description: user_id
          in: path
          name: user_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/userModel.UserShow"
      summary: Menampilkan detail user
      tags:
        - user
swagger: "2.0"